make subsets spec less picky
The ordering of elements within a subset should not be considered to be
significant. So a subset of [1, 2] should be considered equivalent to a
subset of [2, 1]. By sorting each of the user's subsets, and by
providing an expected set of subsets where the subsets are each sorted,
the spec is made to be less sensitive to inconsequential implementation
details.