Make common_subsets spec independent of subset ordering
The ordering of elements within a subset should not be considered to be
significant, e.g. a subset of [1, 2] should be considered equivalent to
a subset of [2, 1]. By sorting each of the user's returned subsets, and
by providing an expected list of subsets where the elements of the
subsets are sorted, the spec is made to be less sensitive to
inconsequential implementation details.

Similarly, the ordering of the subsets should not be considered to be
significant, e.g. `[[], [5]]` and `[[5], []]` should both pass; this
change accomplishes that by using `match_array` rather than `eq`.